refactored duplicate styles into baselayout and cleaned up several redundancies
