+++ client/src/config/constants.ts
@@ -1,0 +1,4 @@
+export const constants = {
+    bugReportUrl: "https://github.com/Propo41/bookify/issues/new?assignees=&labels=bug&projects=&template=bug_report.md&title=%5BBUG%5D",
+    featureRequestUrl: "https://github.com/Propo41/bookify/issues/new?assignees=&labels=feature+request&projects=&template=feature_request.md&title=%5BFEATURE+REQ%5D"
+}
